+++ src/Resources/app/storefront/src/sas-maps/sas-maps.plugin.js
@@ -2,6 +2,9 @@ import Plugin from 'src/plugin-system/plugin.class';
 import DomAccess from 'src/helper/dom-access.helper';
 
 import mapboxgl from 'mapbox-gl';
+import MapboxGeocoder from '@mapbox/mapbox-gl-geocoder';
+import MapboxLanguage from '@mapbox/mapbox-gl-language';
+
 
 export default class Maps extends Plugin {
 
@@ -15,6 +18,7 @@ export default class Maps extends Plugin {
         const description = DomAccess.getDataAttribute(this.el, 'description');
         const mapboxStyle = DomAccess.getDataAttribute(this.el, 'mapbox');
         const zoom = DomAccess.getDataAttribute(this.el, 'zoom');
+        const language = DomAccess.getDataAttribute(this.el, 'language');
         const apiKey = DomAccess.getDataAttribute(this.el, 'token');
 
         mapboxgl.accessToken = apiKey;
@@ -27,11 +31,19 @@ export default class Maps extends Plugin {
             attributionControl: false
         });
 
+	map.addControl(new MapboxLanguage({
+		defaultLanguage: language
+		}));
+
         new mapboxgl.Marker().setLngLat([geoLat, geoLong]).addTo(map);
 
+	if (description){
         new mapboxgl.Popup({ closeOnClick: false, offset: 40 })
             .setLngLat([geoLat, geoLong])
             .setHTML(description)
             .addTo(map);
+	};
+
+
     }
 }