+++ CHANGELOG.md
@@ -1,3 +1,10 @@
+# 1.0.4
+- added language selection 
+- fixed issue #7 plugin working with Shopware 6.4.13.0
+- fixed issue #6 show Popup Content only if content is set
+- actualised Mapbox codebase versions 
+- changed default map location to London 
+
 # 1.0.3
 - added API key check
 - added a new CMS block